Adding coverage for consumer_utils.py and renaming command_producer.py and status_consumer.py to have kafka prefix
