+++ nexus_constructor/file_writer_ctrl_window.py
@@ -12,8 +12,8 @@ from PySide2.QtCore import QTimer, QAbstractItemModel, QSettings
 from PySide2.QtGui import QStandardItemModel, QCloseEvent, Qt
 from PySide2 import QtCore
 from nexus_constructor.instrument import Instrument
-from nexus_constructor.kafka.status_consumer import StatusConsumer
-from nexus_constructor.kafka.command_producer import CommandProducer
+from nexus_constructor.kafka.kafka_status_consumer import StatusConsumer
+from nexus_constructor.kafka.kafka_command_producer import CommandProducer
 import time
 from nexus_constructor.json.filewriter_json_writer import (
     NexusToDictConverter,
